requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Async task processing
celery==5.3.4
redis==5.0.1

# Machine Learning and Embeddings
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1